requirements.txt
colorama==0.4.6
openai==0.27.2
pandas==1.5.2
python-dotenv==1.0.0
torch==1.12.1
tqdm==4.64.1
